Advertisement edit added - Bug fix user deletion fixed

+++ Wetzelplaats/Wetzelplaats-ejb/src/java/nl/wetzel/entities/Advertisement.java
@@ -184,7 +184,7 @@ public class Advertisement implements Serializable {
         return price;
     }
 
-    public void setPrice(long price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
